@@ -1,15 +1,15 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 
 import PackageDescription
 
 let package = Package(
     name: "SwiftGraphics",
     platforms: [
-        .iOS(.v17),
-        .macOS(.v14),
-        .macCatalyst(.v17),
-        .tvOS(.v17),
-        .visionOS(.v1),
+        .iOS(.v18),
+        .macOS(.v15),
+        .macCatalyst(.v18),
+        .tvOS(.v18),
+        .visionOS(.v2),
     ],
     products: [
         .library(name: "Array2D", targets: ["Array2D"]),
