@@ -1,7 +1,7 @@
 import ModelIO
 
 // TODO: FIXME
-public let oneTrueVertexDescriptor: MDLVertexDescriptor = {
+nonisolated(unsafe) public let oneTrueVertexDescriptor: MDLVertexDescriptor = {
     let vertexDescriptor = MDLVertexDescriptor()
     vertexDescriptor.addOrReplaceAttribute(.init(name: MDLVertexAttributePosition, format: .float3, offset: 0, bufferIndex: 0))
     vertexDescriptor.addOrReplaceAttribute(.init(name: MDLVertexAttributeNormal, format: .float3, offset: 0, bufferIndex: 0))
